Use a generic ini parser as a back end
This splits the code from a pacman-conf specific parser to a generic ini
parser which now acts as a back end for the pacman-conf specific parser.

The idea behind this is that the generic ini parser can be imported on
its own and used for other non related purposes. Specificly tools that
make use of pacman's config while also wanting to create a config file
of their own.